Rename files and remove unused imports
