+++ aquality_selenium_core/configurations/element_cache_configuration.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-from abc import ABC, abstractmethod
+from abc import ABC
 
 
 class ElementCacheConfiguration(ABC):
