+++ aquality_selenium_core/elements/element_state_provider.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-from abc import ABC, abstractmethod
+from abc import ABC
 
 
 class ElementStateProvider(ABC):
