+++ aquality_selenium_core/localization/localized_logger.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-from abc import ABC, abstractmethod
+from abc import ABC
 
 
 class LocalizedLogger(ABC):
