+++ aquality_selenium_core/utilities/element_action_retrier.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-from abc import ABC, abstractmethod
+from abc import ABC
 
 
 class ElementActionRetrier(ABC):
