+++ aquality_selenium_core/waitings/conditional_wait.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-from abc import ABC, abstractmethod
+from abc import ABC
 
 
 class ConditionalWait(ABC):
